+++ src/corpus/analysis/AnalysisPanel.vue
@@ -1,23 +1,30 @@
 <script setup lang="ts">
 import { computed, ref, watch } from "vue";
 import { PhDownloadSimple, PhGearFine, PhInfo } from "@phosphor-icons/vue";
-import useCorpusIdParam from "../corpusIdParam.composable";
-import { useCorpusState } from "../corpusState.composable";
 import useExports from "../exports/exports.composable";
 import useJob from "../job/job.composable";
+import useConfig from "../config/config.composable";
 import ActionButton from "@/components/ActionButton.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 
-const corpusId = useCorpusIdParam();
-const { runJob, jobState, isJobRunning } = useJob(corpusId);
-const { canBeReady } = useCorpusState(corpusId);
+const props = defineProps<{
+  corpusId: string;
+}>();
+
+const { isConfigValid, hasMetadata } = useConfig(props.corpusId);
+const { runJob, clearAnnotations, jobState, isJobRunning } = useJob(
+  props.corpusId,
+);
 const { exports, loadExports, downloadResult, getDownloadFilename } =
-  useExports(corpusId);
-const { isDone } = useCorpusState(corpusId);
+  useExports(props.corpusId);
 
 const isPending = ref(false);
 const canRun = computed(
-  () => canBeReady.value && !isPending.value && !isJobRunning.value,
+  () =>
+    isConfigValid.value &&
+    hasMetadata.value &&
+    !isPending.value &&
+    !isJobRunning.value,
 );
 
 loadExports();
@@ -29,8 +36,8 @@ async function doRunJob() {
 }
 
 // When a job finishes, show download button.
-watch(isDone, () => {
-  if (isDone.value) {
+watch(jobState, () => {
+  if (jobState.value?.sparv == "done") {
     loadExports();
   }
 });
@@ -40,22 +47,40 @@ watch(isDone, () => {
   <div>
     <PendingContent
       :on="`corpus/${corpusId}/job/sparv`"
-      class="flex flex-wrap gap-3 items-start"
+      class="flex flex-col gap-3 items-start"
     >
-      <i18n-t keypath="analysis.help" scope="global" tag="p">
-        <template #sparv>
-          <a :href="$t('analysis.sparv.url')">Sparv</a>
-        </template>
-      </i18n-t>
+      <div v-if="exports?.length" class="flex gap-3 items-center">
+        <div>
+          <div class="font-semibold">{{ $t("annotations.clear") }}</div>
+          {{ $t("annotations.clear.help") }}
+        </div>
+        <ActionButton @click="clearAnnotations()" class="whitespace-nowrap">
+          {{ $t("annotations.clear") }}
+        </ActionButton>
+      </div>
 
-      <ActionButton
-        :disabled="isJobRunning || !canRun"
-        :class="{ 'button-primary': !exports?.length }"
-        @click="!isJobRunning && canRun ? doRunJob() : null"
-      >
-        <PhGearFine weight="bold" class="inline mb-1 mr-1" />
-        {{ !exports?.length ? $t("job.run") : $t("job.rerun") }}
-      </ActionButton>
+      <div class="flex gap-3 items-center">
+        <div>
+          <div class="font-semibold">{{ $t("job.run") }}</div>
+          <i18n-t keypath="analysis.help" scope="global">
+            <template #sparv>
+              <a :href="$t('analysis.sparv.url')">Sparv</a>
+            </template>
+          </i18n-t>
+        </div>
+
+        <ActionButton
+          :disabled="isJobRunning || !canRun"
+          :class="{
+            'button-primary': !isJobRunning && canRun && !exports?.length,
+          }"
+          @click="!isJobRunning && canRun ? doRunJob() : null"
+          class="whitespace-nowrap"
+        >
+          <PhGearFine weight="bold" class="inline mb-1 mr-1" />
+          {{ !exports?.length ? $t("job.run") : $t("job.rerun") }}
+        </ActionButton>
+      </div>
 
       <div>
         <div v-if="!isJobRunning && exports?.length" class="text-sm">