+++ src/corpus/config/config.composable.ts
@@ -2,10 +2,12 @@ import { computed } from "vue";
 import {
   makeConfig,
   parseConfig,
+  validateConfig,
   type ConfigOptions,
 } from "@/api/corpusConfig";
 import useLocale from "@/i18n/locale.composable";
 import { useResourceStore } from "@/store/resource.store";
+import { getException } from "@/util";
 
 export default function useConfig(corpusId: string) {
   const resourceStore = useResourceStore();
@@ -15,6 +17,14 @@ export default function useConfig(corpusId: string) {
   const config = computed(() => corpus.value?.config);
   const configOptions = computed(getParsedConfig);
   const corpusName = computed(() => th(configOptions.value?.name));
+  const hasMetadata = computed(
+    () => configOptions.value?.name?.swe || configOptions.value?.name?.eng,
+  );
+  const isConfigValid = computed(
+    () =>
+      configOptions.value &&
+      !getException(() => validateConfig(configOptions.value!)),
+  );
 
   async function uploadConfig(configOptions: ConfigOptions) {
     const configYaml = makeConfig(corpusId, configOptions);
@@ -35,6 +45,8 @@ export default function useConfig(corpusId: string) {
     config,
     configOptions,
     corpusName,
+    hasMetadata,
+    isConfigValid,
     uploadConfig,
   };
 }